debug: Track entity id per render resource

+++ libs/rend/include/rend_resource.h
@@ -17,7 +17,7 @@ bool  rend_res_is_loading(const RendResComp*);
 bool  rend_res_is_failed(const RendResComp*);
 bool  rend_res_is_unused(const RendResComp*);
 bool  rend_res_is_persistent(const RendResComp*);
-u64   rend_res_ticks_until_unload(const RendResComp*);
+u32   rend_res_ticks_until_unload(const RendResComp*);
 usize rend_res_mesh_data_size(const RendResMeshComp*);
 usize rend_res_texture_data_size(const RendResTextureComp*);
